test: isolate Sheet comparison case in its own file
